% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilizeHTSLIB.R
\name{findVcfFiles}
\alias{findVcfFiles}
\title{Get all paths to the .vcf and .vcf.gz files in the given directories}
\usage{
findVcfFiles(dirs)
}
\arguments{
\item{dirs}{an array of directories to search the .vcf and .vcf.gz files}
}
\value{
an array of all paths to the .vcf and .vcf.gz files in the directories
}
\description{
This function reads an array of directories and returns an array of all paths to the .vcf and .vcf.gz files in the directories
}
