% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilizeHTSLIB.R
\name{readSNPonlyDSwithPositionVector}
\alias{readSNPonlyDSwithPositionVector}
\title{Read SNP's Dosage with position array and index}
\usage{
readSNPonlyDSwithPositionVector(filename, positions, index)
}
\arguments{
\item{positions}{an array of positions for each SNP in the bd file}

\item{index}{the index for the SNP in the file}

\item{file}{a bd file to read the SNP}
}
\value{
an array of Dosage data
}
\description{
This function reads an array of Dosage data, given an array of positions for each SNP in the bd file and the index for the SNP in the file
}
