% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilizeHTSLIB.R
\name{readSNPwithPositionVector}
\alias{readSNPwithPositionVector}
\title{Read SNP with position array and index}
\usage{
readSNPwithPositionVector(filename, positions, index)
}
\arguments{
\item{positions}{an array of positions for each SNP in the bd file}

\item{index}{the index for the SNP in the file}

\item{file}{a bd file to read the SNP}
}
\value{
a dataframe of sample ID, DS and GP
}
\description{
This function reads a dataframe of sample ID, DS and GP, given an array of positions for each SNP in the bd file and the index for the SNP in the file
}
